backend pushed

+++ src/main/kotlin/com/sorsix/fitness/service/UserService.kt
@@ -13,7 +13,9 @@ import com.sorsix.fitness.repository.PostRepository
 import com.sorsix.fitness.repository.UserFollowUserRepository
 import com.sorsix.fitness.repository.UserLikePostRepository
 import com.sorsix.fitness.repository.UserRepository
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Service
+import org.springframework.web.bind.annotation.PathVariable
 import java.util.*
 import javax.transaction.Transactional
 
@@ -27,39 +29,49 @@ class UserService(
     val encoder: PasswordEncoder) {
 
     @Transactional
-    fun followUser(followingId: Long, followerId: Long): Boolean {
-        val userFollowing: Optional<User> = userRepository.findById(followingId)
+    fun followUser(followingId: Long, followerId: Long): Response<*> {
+        val userFollowing:User = SecurityContextHolder.getContext().authentication.principal as User
         val userFollower: Optional<User> = userRepository.findById(followerId)
+
+        if(userFollower.isEmpty){
+            return NotFound("User with id $followerId was not found")
+        }
 
         return if(userFollowUserRepository.existsByUserFollowingIdAndUserFollowerId(followingId,followerId)){
             userFollowUserRepository.deleteUserFollowUserByUserFollowerIdAndUserFollowingId(followerId,followingId)
             userRepository.updateNumberOfFollowersMinus(followerId)
             userRepository.updateNumberOfFollowingMinus(followingId)
             userRepository.save(userFollower.get())
-            userRepository.save(userFollowing.get())
-            false
+            userRepository.save(userFollowing)
+            Success("Ok")
         } else {
-            userFollowUserRepository.save(UserFollowUser(0L,userFollowing.get(),userFollower.get()))
+            userFollowUserRepository.save(UserFollowUser(0L,userFollowing,userFollower.get()))
             userRepository.updateNumberOfFollowersPlus(followerId)
             userRepository.updateNumberOfFollowingPlus(followingId)
             userRepository.save(userFollower.get())
-            userRepository.save(userFollowing.get())
-            true
+            userRepository.save(userFollowing)
+            Success("Ok")
         }
     }
 
     @Transactional
-    fun musclePost(userId: Long, postId: Long): Boolean {
-        return if(userLikePostRepository.existsByUserIdAndPostId(userId,postId)){
+    fun musclePost(@PathVariable postId: Long): Response<*> {
+        val user = SecurityContextHolder.getContext().authentication.principal as User
+
+        val post: Optional<Post> = this.postRepository.findById(postId)
+
+        if (post.isEmpty) {
+            return NotFound("Post with id $postId was not found")
+        }
+
+        return if(userLikePostRepository.existsByUserIdAndPostId(user.id, postId)) {
             postRepository.dislike(postId)
-            userLikePostRepository.deleteUserLikePostByUserIdAndPostId(userId,postId)
-            false
+            userLikePostRepository.deleteUserLikePostByUserIdAndPostId(user.id, postId)
+            Success("Post disliked successfully")
         } else {
             postRepository.like(postId)
-            val user: Optional<User> = userRepository.findById(userId)
-            val post: Optional<Post> = postRepository.findById(postId)
-            userLikePostRepository.save(UserLikePost(0L,user.get(),post.get()))
-            true
+            userLikePostRepository.save(UserLikePost(0L, user, post.get()))
+            Success("Post liked successfully")
         }
     }
